[RS_A2_04]: renamed container to card
