--- src/components/Container/index.ts
@@ -1,1 +1,0 @@
-export * from './Container.component';
